// src/lib/api/generated/models/domainBrewRecipeResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Coffeeroasters API
 * OpenAPI spec version: 1.0
 */

export interface DomainBrewRecipeResponse {
  id?: number;
  coffee_id?: number;
  user_id?: number;
  user_email?: string;
  title?: string;
  brew_method?: string;
  dose_grams?: number;
  water_ml?: number;
  water_temp_c?: number;
  grind_size?: string;
  brew_time_seconds?: number;
  notes?: string;
  is_public?: boolean;
  coffee_name?: string;
  roaster_name?: string;
  roaster_slug?: string;
  created_at?: string;
  updated_at?: string;
}